backend/requirements.txt
# Custom ERP Backend
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
marshmallow>=3.20.1
marshmallow-sqlalchemy>=0.29.0
authlib>=1.3.0
requests>=2.31.0
gunicorn>=21.0.0
